refactor #26 - CustomShapeView를 분리하여 셀 UI 개선 - UICollectionViewCell 내부에서 layoutSubviews()를 사용하지 않고 CustomView로 분리 - CustomView에서 mask와 gradient 적용을 담당하도록 변경 - 초기 로딩 시에도 올바른 모양과 색상이 적용되도록 수정
